CRUD users associate with permissions

+++ application/helpers/auth_helper.php
@@ -7,7 +7,7 @@ function auth_user(){
 }
 
 function auth_id(){
-	return Auth::id();
+	return (int) Auth::id();
 }
 
 function auth_check(){
